move NextQueueItemIsValid impl
